Refactor: #103 - place 대신 room 필수 입력으로 변경
변경:
createItem2()
itemToItemRS2()
itemToItemRSList2()
updateItem2()
deleteLocation2()

변경 X:
getItemsInLocation()

+++ src/main/java/com/house/item/service/LocationService.java
@@ -137,7 +137,30 @@ public class LocationService {
         }
 
         if (!items.isEmpty()) {
-            throw new UnableToDeleteLocationInUseException(ExceptionCodeMessage.UNABLE_TO_DELETE_LOCATION_IN_USE_EXCEPTION.message());
+            throw new UnableToDeleteLocationInUseException(
+                ExceptionCodeMessage.UNABLE_TO_DELETE_LOCATION_IN_USE_EXCEPTION.message());
+        }
+
+        locationRepository.delete(location);
+    }
+
+    @Transactional
+    public void deleteLocation2(Long locationNo, User user) {
+        Location location = getLocation(locationNo, user);
+
+        List<Item> items = itemRepository.findByLocation(location);
+        if (location.getType() == LocationType.ROOM) {
+            List<Location> placeList = locationRepository.findByRoom(location);
+            items.addAll(itemRepository.findByLocationIn(placeList));
+        }
+
+        if (items == null) {
+            throw new UndefinedLocationTypeException(ExceptionCodeMessage.UNDEFINED_LOCATION_TYPE.message());
+        }
+
+        if (!items.isEmpty()) {
+            throw new UnableToDeleteLocationInUseException(
+                ExceptionCodeMessage.UNABLE_TO_DELETE_LOCATION_IN_USE_EXCEPTION.message());
         }
 
         locationRepository.delete(location);
